- move common code into separate file
